Move accordion data to constants and add EDD variant to confirmation
