+++ pages/hold/confirmation/[id].tsx
@@ -2,22 +2,32 @@ import Head from "next/head"
 import {
   Banner,
   Accordion,
-  type AccordionDataProps,
   Heading,
   Text,
 } from "@nypl/design-system-react-components"
 
 import Layout from "../../../src/components/Layout/Layout"
-import { SITE_NAME } from "../../../src/config/constants"
-import RCLink from "../../../src/components/Links/RCLink/RCLink"
-import ExternalLink from "../../../src/components/Links/ExternalLink/ExternalLink"
+import {
+  SITE_NAME,
+  HOLD_PAGE_HEADING,
+  EDD_PAGE_HEADING,
+  HOLD_CONFIRMATION_FAQ_DATA,
+  EDD_CONFIRMATION_FAQ_DATA,
+} from "../../../src/config/constants"
+
 // import HoldItemDetails from "../../../src/components/HoldPages/HoldItemDetails"
+
+interface HoldConfirmationPageProps {
+  isEDD?: boolean
+}
 
 /**
  * The Hold Confirmation page shows a success message, item details, and a FAQ accordion
  * when a patron successfully places a hold request.
  */
-export default function HoldConfirmationPage() {
+export default function HoldConfirmationPage({
+  isEDD = false,
+}: HoldConfirmationPageProps) {
   const metadataTitle = `Request Confirmation | ${SITE_NAME}`
   return (
     <>
@@ -33,7 +43,7 @@ export default function HoldConfirmationPage() {
       </Head>
       <Layout activePage="hold">
         <Heading level="h2" mb="l" size="heading3">
-          Request for on-site use
+          {isEDD ? EDD_PAGE_HEADING : HOLD_PAGE_HEADING}
         </Heading>
         {/* <HoldItemDetails item={item} /> */}
         <Banner
@@ -50,93 +60,13 @@ export default function HoldConfirmationPage() {
         <Heading level="h3" mb="l">
           Frequently asked questions
         </Heading>
-        <Accordion accordionData={faqContentData} isDefaultOpen />
+        <Accordion
+          accordionData={
+            isEDD ? EDD_CONFIRMATION_FAQ_DATA : HOLD_CONFIRMATION_FAQ_DATA
+          }
+          isDefaultOpen
+        />
       </Layout>
     </>
   )
 }
-
-const faqContentData: AccordionDataProps[] = [
-  {
-    label: "What's next?",
-    panel: (
-      <>
-        <Text>
-          Please allow a few minutes for this item to show up in your{" "}
-          <RCLink href="/account">patron account</RCLink>.
-        </Text>
-        <Text>
-          The item will be listed as &quot;Ready for pickup&quot; under your
-          holds tab when it is available. You will receive an email confirmation
-          after your item has arrived.
-        </Text>
-      </>
-    ),
-  },
-  {
-    label: "When will my item be ready for pickup?",
-    panel: (
-      <>
-        <Text>
-          <strong>Items stored on-site:</strong> Materials requested up to an
-          hour before closing are usually ready for pickup within an hour.
-          Materials requested within an hour of closing or outside business
-          hours are ready about an hour after opening on the next business day.
-        </Text>
-        <Text>
-          <strong>Items stored off-site:</strong> Materials requested before
-          2:30 PM are usually ready for pickup about an hour after opening the
-          next day (check{" "}
-          <ExternalLink href="https://www.nypl.org/">nypl.org</ExternalLink> for
-          library hours). Materials requested after 2:30 PM Mon–Thu are usually
-          ready in two days; materials requested after 2:30 PM Fri–Sun are ready
-          on Tuesday. Some materials are not delivered on Saturdays.
-        </Text>
-      </>
-    ),
-  },
-  {
-    label: "How long will my item be available for?",
-    panel: (
-      <>
-        <Text>
-          We will hold books for up to 14 days, so you can request materials up
-          to two weeks in advance.
-        </Text>
-      </>
-    ),
-  },
-  {
-    label: "How do I pick up my item once it is ready?",
-    panel: (
-      <>
-        <Text>
-          Once your item is ready for pickup, please arrive at the pickup
-          location during business hours and proceed to a help desk. An NYPL
-          staff member will check your item out to you.
-        </Text>
-      </>
-    ),
-  },
-  {
-    label: "How do I cancel my request?",
-    panel: (
-      <>
-        <Text>
-          If you would like to cancel your request or have questions, please{" "}
-          <ExternalLink href="https://gethelp.nypl.org/customer/portal/emails/new">
-            email us
-          </ExternalLink>{" "}
-          or call 917-ASK-NYPL (
-          <ExternalLink href="tel:19172756975">917-275-6975</ExternalLink>).
-          Processed requests can also be canceled from the holds tab in your
-          patron account.
-        </Text>
-        <Text>
-          For more information about our requesting services, please see
-          Requesting Research Materials.
-        </Text>
-      </>
-    ),
-  },
-]
